create 1 round play and use it for multiple games
I have created a function which uses computer play generator to
compare it with user input and determine a 1 round winner.
I have then used this function to create a game function which
plays 5 rounds. First to get 3 wins is a final winner.
Both of the functions take into account incorrect input or a tie.